[BGP] Fix FRRConfiguration cleanup
Currently the process of processing the FRRConfigurations is:
* get a list of all pods in the ctlplane namespace
* filter that list to the pods which have a secondary IF configured
* run through that list of pods with an IF attached and validate
  if there is a change of the node it is running on.

While running through that nested loop to check if there is
a an update needed we are also checking there is an frr config
for an already deleted pod.

This is ok unless there is at least one pod with a secondary IF, if
not the cleanup won't happen, which is the case for the FRRConfiguration
for the last deleted pod.

This issue was only seen randomly because the functional test
was not waiting for the pod and its FRRConfiguration to exist
before it started the delete test.

This change fixes the functional test and moves the cleanup out of
the mentioned loop so it also gets checked properly when the
last pod gets deleted.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>